requirements.txt
2captcha-python==1.0.3
aiodns==3.0.0
anime-downloader==5.0.12
anticaptchaofficial==1.0.30
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bestcaptchasolverapi2==0.2
bestcaptchasolverapi3==0.2
browser-cookie3==0.12.1
bs4==0.0.1
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.5
cfscrape==2.1.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cloudscraper==1.2.58
coinbase==2.1.0
colorama==0.4.4
coloredlogs==15.0.1
colorgram.py==1.2.0
cryptography==3.4.7
deathbycaptcha==0.1.1
decorator==4.4.2
Django==3.1.6
django-request==1.5.6
fake-useragent==0.1.11
Flask==2.0.2
Flask-SQLAlchemy==2.4.4
fuzzywuzzy==0.18.0
greenlet==1.1.2
html5lib==1.1
humanfriendly==9.2
idna==3.2
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==2.0.1
jedi==0.17.2
jeepney==0.6.0
Jinja2==3.0.2
jupyter-client==6.1.7
jupyter-core==4.6.3
keyring==23.0.1
lazy-object-proxy==1.4.3
lz4==3.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
MouseInfo==0.1.3
multidict==5.1.0
multithreading==0.2.0
numpy==1.20.0
opencv-python==4.5.1.48
packaging==21.0
pandas==1.2.1
parso==0.7.1
pbkdf2==1.3
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.8
py==1.10.0
pyaes==1.6.1
PyAutoGUI==0.9.52
pycares==4.0.0
pycparser==2.20
pycryptodome==3.10.1
PyGetWindow==0.0.9
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
PyMouse==1.0
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyreadline==2.1
PyRect==0.1.4
PyScreeze==0.1.26
PySimpleGUI==4.41.2
pySmartDL==1.3.4
pytest==6.2.4
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.2
PyTweening==1.0.3
pytz==2021.1
pyuseragents==1.0.5
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==19.0.2
rawgpy==1.0.3
requests==2.26.0
requests-cache==0.7.3
requests-toolbelt==0.9.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
SecretStorage==3.3.1
selenium==3.141.0
shadow-useragent==0.0.17
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
stockx-py-sdk==0.0.2
tabulate==0.8.9
tmdbv3api==1.7.4
toml==0.10.2
tornado==6.0.4
traitlets==5.0.5
twilio==6.52.0
twitter==1.19.3
TwoCaptcha==0.0.1
typing-extensions==3.10.0.0
url-normalize==1.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
